+++ crates/tytanic/src/cli/commands/run.rs
@@ -34,10 +34,27 @@ pub fn run(ctx: &mut Context, args: &Args) -> eyre::Result<()> {
     let suite = ctx.collect_tests_with_filter(&project, ctx.filter(&args.filter)?)?;
     let world = ctx.world(&args.compile)?;
 
-    let origin = match args.export.dir {
+    let origin = match args
+        .export
+        .dir
+        .map(OptionDelegate::into_native)
+        .unwrap_or(project.config().defaults.direction)
+    {
         Direction::Ltr => Origin::TopLeft,
         Direction::Rtl => Origin::TopRight,
     };
+
+    let pixel_per_pt = render::ppi_to_ppp(args.export.ppi.unwrap_or(project.config().defaults.ppi));
+
+    let max_delta = args
+        .compare
+        .max_delta
+        .unwrap_or(project.config().defaults.max_delta);
+
+    let max_deviation = args
+        .compare
+        .max_deviations
+        .unwrap_or(project.config().defaults.max_deviations);
 
     let runner = Runner::new(
         &project,
@@ -47,15 +64,15 @@ pub fn run(ctx: &mut Context, args: &Args) -> eyre::Result<()> {
             warnings: args.compile.warnings.into_native(),
             optimize: args.export.optimize_refs.get_or_default(),
             fail_fast: args.runner.fail_fast.get_or_default(),
-            pixel_per_pt: render::ppi_to_ppp(args.export.ppi),
+            pixel_per_pt,
             action: Action::Run {
                 strategy: args
                     .compare
                     .compare
                     .get_or_default()
                     .then_some(Strategy::Simple {
-                        max_delta: args.compare.max_delta,
-                        max_deviation: args.compare.max_deviations,
+                        max_delta,
+                        max_deviation,
                     }),
                 export_ephemeral: args.export.export_ephemeral.get_or_default(),
                 origin,